+++ Realm/Realm/Sync/FlexibleSync/SubscriptionSet.cs
@@ -21,6 +21,7 @@ using System.Collections;
 using System.Collections.Generic;
 using System.Diagnostics.CodeAnalysis;
 using System.Linq;
+using System.Threading;
 using System.Threading.Tasks;
 using Realms.Helpers;
 using Realms.Sync.Exceptions;
@@ -296,6 +297,7 @@ namespace Realms.Sync
         /// <summary>
         /// Waits for the server to acknowledge the subscription set and return the matching objects.
         /// </summary>
+        /// <param name="cancellationToken">An optional cancellation token that can be used to cancel the wait operation.</param>
         /// <remarks>
         /// If the <see cref="State"/> of the subscription set is <see cref="SubscriptionSetState.Complete"/>
         /// the returned <see cref="Task"/> will complete immediately. If the <see cref="State"/> is
@@ -309,7 +311,7 @@ namespace Realms.Sync
         /// An awaitable task, whose successful completion indicates that the server has processed the
         /// subscription change and has sent all the data that matches the new subscriptions.
         /// </returns>
-        public Task WaitForSynchronizationAsync() => _handle.WaitForStateChangeAsync();
+        public Task WaitForSynchronizationAsync(CancellationToken? cancellationToken = null) => _handle.WaitForStateChangeAsync(cancellationToken);
 
         /// <inheritdoc/>
         public IEnumerator<Subscription> GetEnumerator() => new Enumerator(this);